add file uploader and convert image to base 64

+++ src/dto/request/pet_request.go
@@ -9,5 +9,5 @@ type PetRequest struct {
 	Gender      string  `json:"gender" validate:"len=1"`
 	Color       string  `json:"color" validate:"required"`
 	Weight      float32 `json:"weight" validate:"required"`
-	UserID      uint    `json:"user_id" validate:"required"`
+	UserID      uint64  `json:"user_id" validate:"required"`
 }
